Refactor(fem): restructure interface poisson model code to conform to standard.
