+++ fealpy/mesher/__init__.py
@@ -6,4 +6,5 @@ from .cylinder_mesher import CylinderMesher
 from .interval_mesher import IntervalMesher
 from .sphere_shell_mesher import SphereShellMesher
 from .hollowblock_mesher import HollowBlockMesher
+from .interface_mesher import InterfaceFittedMesher2d
