+++ fealpy/model/interface_poisson/__init__.py
@@ -19,10 +19,9 @@ class InterfacePoissonDataProtocol(Protocol):
     """
     def geo_dimension(self) -> int: ...
     def domain(self) -> Sequence[float]: ...
-    def level_function(p) -> TensorLike: ...
     def solution(self, p: TensorLike) -> TensorLike: ...
     def gradient(self, p: TensorLike) -> TensorLike: ...
-    def gN(self, p: TensorLike) -> TensorLike: ...
+    def flux_interface_condition(self, p: TensorLike) -> TensorLike: ...
     def source(self, p: TensorLike) -> TensorLike: ...
     def dirichlet(self, p: TensorLike) -> TensorLike: ...
     def is_dirichlet_boundary(self, p: TensorLike) -> TensorLike: ...
@@ -35,5 +34,5 @@ follow the existing examples to register them in the registry.
 """
 DATA_TABLE = {
     # example name: (file_name, class_name)
-    "electronic": ("electronic_data_2d", "ElectronicData2D")
+    "1": ("exp0001", "Exp0001")
 }
